Split shop mutations file
