+++ saleor/graphql/shop/mutations/__init__.py
@@ -1,0 +1,21 @@
+from .gift_card_settings_update import GiftCardSettingsUpdate
+from .order_settings_update import OrderSettingsUpdate
+from .shop_address_update import ShopAddressUpdate
+from .shop_domain_update import ShopDomainUpdate
+from .shop_fetch_tax_rates import ShopFetchTaxRates
+from .shop_settings_update import ShopSettingsUpdate
+from .staff_notification_recipient_create import StaffNotificationRecipientCreate
+from .staff_notification_recipient_delete import StaffNotificationRecipientDelete
+from .staff_notification_recipient_update import StaffNotificationRecipientUpdate
+
+__all__ = [
+    "GiftCardSettingsUpdate",
+    "OrderSettingsUpdate",
+    "ShopAddressUpdate",
+    "ShopDomainUpdate",
+    "ShopFetchTaxRates",
+    "ShopSettingsUpdate",
+    "StaffNotificationRecipientCreate",
+    "StaffNotificationRecipientDelete",
+    "StaffNotificationRecipientUpdate",
+]
